refactor: 여백 수정 및 컴포넌트 분리
