+++ frontend/app/src/main/java/com/apptive/japkor/ui/components/step4/PhotoUploadButton.kt
@@ -1,0 +1,52 @@
+package com.apptive.japkor.ui.components.step4
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun PhotoUploadButton(
+    modifier: Modifier = Modifier,
+    onClick: () -> Unit = {}
+) {
+    val shape = RoundedCornerShape(10.dp)
+
+    // 바탕 카드 (soft shadow)
+    Box(
+        modifier = modifier
+            .shadow(2.dp, shape, clip = false)
+            .clip(shape)
+            .background(Color.White)
+            .clickable(onClick = onClick),
+        contentAlignment = Alignment.Center
+    ) {
+        // 가운데 옅은 원 + '+' 아이콘
+        Box(
+            modifier = Modifier
+                .size(25.dp)                          // 원 크기
+                .background(
+                    color = Color(0xFFF5F6F8),        // 연한 회색 배경
+                    shape = CircleShape
+                ),
+            contentAlignment = Alignment.Center
+        ) {
+            Text(
+                text = "+",
+                fontSize = 22.sp,
+                color = Color(0xFFB8BDC7)            // 아이콘 컬러
+            )
+        }
+    }
+}
