+++ frontend/app/src/main/java/com/apptive/japkor/ui/requiredinfo/steps/Step4Content.kt
@@ -1,31 +1,21 @@
 package com.apptive.japkor.ui.requiredinfo.steps
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.apptive.japkor.ui.components.CustomText
 import com.apptive.japkor.ui.components.CustomTextType
+import com.apptive.japkor.ui.components.step4.PhotoUploadGrid
 import com.apptive.japkor.ui.theme.CustomColor
 
 @Composable
@@ -51,13 +41,14 @@ fun Step4Content() {
 
         PhotoUploadGrid()
 
-        Spacer(modifier = Modifier.height(5.dp))
+
 
         CustomText(
             text = "- 사진은 프로필에서 가장 중요한 요소입니다.",
             color = CustomColor.gray300,
             type = CustomTextType.bodyLarge,
-            modifier = Modifier.padding(horizontal = 7.dp)
+            modifier = Modifier.padding(horizontal = 7.dp),
+            size = 16.sp
         )
 
         CustomText(
@@ -74,7 +65,7 @@ fun Step4Content() {
             modifier = Modifier.padding(horizontal = 7.dp)
         )
 
-        Spacer(modifier = Modifier.height(85.dp))
+
         CustomText(
             text = "프로필 사진 가이드 보기",
             color = CustomColor.gray300,
@@ -92,63 +83,6 @@ fun Step4Content() {
 }
 
 // 3열 그리드 (간격/패딩/정사각형 유지)
-@Composable
-fun PhotoUploadGrid() {
-    Column(modifier = Modifier.fillMaxWidth()) {
-        val cells = (0 until 6).toList()
-        Column(verticalArrangement = Arrangement.spacedBy(10.dp)) {
-            cells.chunked(3).forEach { row ->
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(10.dp)
-                ) {
-                    row.forEach { idx ->
-                        UploadPhotoCard(
-                            modifier = Modifier
-                                .weight(1f)
-                                .aspectRatio(1f),   // 정사각형
-                            onClick = { /* TODO: photo pick for idx */ }
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
+
 
 // 카드 1칸 (화이트 카드 + 소프트 섀도우 + 중앙 원형 + 아이콘)
-@Composable
-fun UploadPhotoCard(
-    modifier: Modifier = Modifier,
-    onClick: () -> Unit = {}
-) {
-    val shape = RoundedCornerShape(10.dp)
-
-    // 바탕 카드 (soft shadow)
-    Box(
-        modifier = modifier
-            .shadow(2.dp, shape, clip = false)
-            .clip(shape)
-            .background(Color.White)
-            .clickable(onClick = onClick),
-        contentAlignment = Alignment.Center
-    ) {
-        // 가운데 옅은 원 + '+' 아이콘
-        Box(
-            modifier = Modifier
-                .size(25.dp)                          // 원 크기
-                .background(
-                    color = Color(0xFFF5F6F8),        // 연한 회색 배경
-                    shape = CircleShape
-                ),
-            contentAlignment = Alignment.Center
-        ) {
-            Text(
-                text = "+",
-                fontSize = 22.sp,
-                color = Color(0xFFB8BDC7)            // 아이콘 컬러
-            )
-        }
-    }
-}
